+++ src/lib/api/MusicService.ts
@@ -1,5 +1,22 @@
-import { MusicReleaseEntity } from "Entity/Music";
 import AxiosInstance from "./AxiosInstance";
+
+export interface MusicEntity {
+  name: string;
+  artist: string;
+  album: string;
+  albumimg: string;
+  date: string;
+  melonlink: string;
+  kakaomelonlink: string;
+  lyrics: string;
+}
+
+export interface MusicReleaseEntity {
+  type: string;
+  status: string;
+  lineup: string[];
+  song: MusicEntity[];
+}
 
 class MusicService {
   static search(keyword: string) {
